Remove Subproject Leads. (#9)
The original intent for subproject leads was to prevent decision paralysis in important projects, and to make sure that there was always someone available in major projects, who had the knowledge, breadth and depth of technical expertise, and insight into their given project to be able to make decisions for the good of the project long into the future.

However, that isn't working. Subproject leads are repeatedly the subject of much debate, and are compared to dictators, despite this being the antithesis of the intent.

As such, they should be removed.

All arguments within project teams should be resolved by the Steering Council - in the absence of the Subproject Leaders, this is a role they fill by default, as outlined in the remaining document.